+++ account_multicurrency_revaluation/__manifest__.py
@@ -1,6 +1,3 @@
-# Copyright 2012-2018 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 {
     "name": "Multicurrency revaluation",
     "version": "17.0.1.0.0",
@@ -10,19 +7,26 @@
     "website": "https://github.com/OCA/account-closing",
     "license": "AGPL-3",
     "depends": ["account"],
-    "demo": ["demo/account_demo.xml", "demo/currency_demo.xml"],
+    "demo": [
+        "demo/res_partner.xml",
+        "demo/res_company.xml",
+        "demo/res_currency_rate.xml",
+        "demo/account_account.xml",
+        "demo/account_journal.xml",
+        "demo/account_analytic.xml",
+    ],
     "data": [
-        "views/res_config_view.xml",
-        "security/security.xml",
+        "views/res_config_settings.xml",
+        "security/account_multicurrency_revaluation_groups.xml",
         "security/ir.model.access.csv",
-        "views/account_view.xml",
-        "views/account_move_views.xml",
-        "views/account_move_line_views.xml",
+        "views/account_account.xml",
+        "views/account_move.xml",
+        "views/account_move_line.xml",
         "wizard/print_currency_unrealized_report_view.xml",
         "wizard/wizard_currency_revaluation_view.xml",
         "wizard/wizard_reverse_currency_revaluation_view.xml",
-        "report/report.xml",
-        "report/unrealized_currency_gain_loss.xml",
+        "report/currency_unrealized_report.xml",
+        "templates/currency_unrealized_report.xml",
     ],
     "assets": {
         "web.report_assets_common": [
