+++ account_multicurrency_revaluation/demo/account_analytic.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record id="analytic_plan_revaluation" model="account.analytic.plan">
+        <field name="name">Revaluation</field>
+    </record>
+    <record id="acc_analytic" model="account.analytic.account">
+        <field name="name">Analytic Account</field>
+        <field name="code">XA001</field>
+        <field name="company_id" ref="res_company_reval" />
+        <field
+            name="plan_id"
+            ref="account_multicurrency_revaluation.analytic_plan_revaluation"
+        />
+    </record>
+</odoo>
