+++ account_multicurrency_revaluation/model/__init__.py
@@ -1,3 +1,5 @@
 from . import res_company
-from . import account
-from . import res_config
+from . import account_account
+from . import account_move
+from . import account_move_line
+from . import res_config_settings
