+++ account_multicurrency_revaluation/model/res_config_settings.py
@@ -1,7 +1,3 @@
-# Copyright 2012-2018 Camptocamp SA
-# Copyright 2022 ForgeFlow S.L. (https://www.forgeflow.com)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from odoo import fields, models
