+++ account_multicurrency_revaluation/tests/test_account_account.py
@@ -1,6 +1,3 @@
-# Copyright 2020 CorporateHub (https://corporatehub.eu)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
 from odoo import Command
 from odoo.tests import common
