+++ account_multicurrency_revaluation/tests/test_currency_revaluation.py
@@ -1,7 +1,3 @@
-# Copyright 2012-2018 Camptocamp SA
-# Copyright 2020 CorporateHub (https://corporatehub.eu)
-# Copyright 2020 Tecnativa - Pedro M. Baeza
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from datetime import timedelta
 
 from odoo import Command, exceptions, fields
