+++ account_multicurrency_revaluation/tests/test_currency_revaluation_report.py
@@ -1,6 +1,3 @@
-# Copyright 2012-2018 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from odoo.tests.common import TransactionCase
 
 from .common import CURRENT_MODULE
