+++ account_multicurrency_revaluation/wizard/wizard_reverse_currency_revaluation.py
@@ -1,6 +1,3 @@
-# Copyright 2022 ForgeFlow S.L. (https://www.forgeflow.com)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from odoo import _, api, exceptions, fields, models
